Split up the MyLeagues file into separate components for better readability.
